move room-list filter options to separate file
