+++ frontend/src/app/dashboard/room-list-filter-options.ts
@@ -1,0 +1,2 @@
+export type FilterOptions = "All" | "Name" | "Location";
+
